converts data type from Float to Double
As over all portfolio value can be large converting to Double.

This helps to fix the limitation of Float

+++ src/main/java/com/learning/mfscreener/models/PortfolioDetailsDTO.java
@@ -3,5 +3,5 @@ package com.learning.mfscreener.models;
 
 import java.io.Serializable;
 
-public record PortfolioDetailsDTO(Float totalValue, String schemeName, String folioNumber, String date)
+public record PortfolioDetailsDTO(Double totalValue, String schemeName, String folioNumber, String date)
         implements Serializable {}
